🚚 Rename : 폴더명 수정 및 구조 변경
