+++ src/main/java/org/toastit_v2/feature/aws/application/service/S3UpLoadService.java
@@ -1,4 +1,4 @@
-package org.toastit_v2.feature.aws.service;
+package org.toastit_v2.feature.aws.application.service;
 
 import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
