+++ src/test/java/org/toastit_v2/feature/aws/service/S3UpLoadServiceImplTest.java
@@ -4,11 +4,11 @@ import com.amazonaws.services.s3.model.ObjectMetadata;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.web.multipart.MultipartFile;
+import org.toastit_v2.feature.aws.application.service.S3UpLoadService;
+import org.toastit_v2.feature.aws.application.service.S3UpLoadServiceImpl;
 
-import java.io.ByteArrayInputStream;
 import java.io.IOException;
 
 import static org.junit.jupiter.api.Assertions.*;
